Modified way of compile.
starting use of babel and webpack for clasp problem that is
import/export.
