+++ package.json
@@ -4,6 +4,8 @@
   "description": "",
   "main": "main.js",
   "scripts": {
+    "build": "webpack",
+    "deploy": "npm build && clasp push",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -17,11 +19,17 @@
   },
   "homepage": "https://github.com/stoikheia/MySmartHome#readme",
   "devDependencies": {
+    "@babel/core": "^7.17.5",
+    "@babel/preset-typescript": "^7.16.7",
     "@types/google-apps-script": "^1.0.45",
+    "babel-loader": "^8.2.3",
+    "gas-webpack-plugin": "^2.1.0",
     "prettier": "^2.5.1",
     "tslint": "^6.1.3",
     "tslint-config-prettier": "^1.18.0",
     "tslint-plugin-prettier": "^2.3.0",
-    "typescript": "^4.5.5"
+    "typescript": "^4.5.5",
+    "webpack": "^5.69.1",
+    "webpack-cli": "^4.9.2"
   }
 }
